create and move search bar from search page to its own file
